+++ azurerm/settings.py
@@ -2,6 +2,7 @@
 
 azure_rm_endpoint = 'https://management.azure.com'
 
+ACS_API = '2016-03-30'
 BASE_API = '2015-01-01'
 COMP_API = '2016-03-30'
 INSIGHTS_API = '2016-03-01'